Inject HttpClient so that benchmarks doesn't have to fetch external sources
