+++ src/HostsParser/ExecutionUtilities.cs
@@ -1,0 +1,125 @@
+// Copyright Henrik Widlund
+// GNU General Public License v3.0
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Globalization;
+using System.IO;
+using System.Net.Http;
+using System.Text;
+using System.Text.Json;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+
+namespace HostsParser;
+
+/// <summary>
+/// Utilities for creating AdBlock files.
+/// </summary>
+public static class ExecutionUtilities
+{
+    /// <summary>
+    /// Reads settings, fetches defined sources and generates a AdBlock file.
+    /// </summary>
+    /// <param name="httpClient">The <see cref="HttpClient"/> that will be used to read from external sources.</param>
+    public static async Task Execute(HttpClient httpClient)
+    {
+        using var loggerFactory = LoggerFactory.Create(options =>
+        {
+            options.AddDebug();
+            options.AddSimpleConsole(consoleOptions =>
+            {
+                consoleOptions.SingleLine = true;
+                consoleOptions.TimestampFormat = CultureInfo.CurrentCulture.DateTimeFormat.ShortDatePattern + " " +
+                                                 CultureInfo.CurrentCulture.DateTimeFormat.LongTimePattern + " ";
+            });
+        });
+        var logger = loggerFactory.CreateLogger(nameof(HostsParser));
+
+        logger.LogInformation("Running...");
+        var stopWatch = new Stopwatch();
+        stopWatch.Start();
+
+        var settings = await JsonSerializer.DeserializeAsync<Settings>(File.OpenRead("appsettings.json"));
+        if (settings == null)
+        {
+            logger.LogError("Couldn't load settings. Terminating...");
+            return;
+        }
+
+        var (combineLines, externalCoverageLines) = await ReadSources(settings, httpClient);
+
+        combineLines.ExceptWith(externalCoverageLines);
+        combineLines = HostUtilities.RemoveKnownBadHosts(settings.KnownBadHosts, combineLines);
+        combineLines.UnionWith(settings.KnownBadHosts);
+        combineLines.UnionWith(externalCoverageLines);
+        CollectionUtilities.FilterGrouped(combineLines);
+
+        var sortedDnsList = CollectionUtilities.SortDnsList(combineLines);
+        HashSet<string> filteredCache = new(combineLines.Count);
+        sortedDnsList = settings.MultiPassFilter
+            ? ProcessingUtilities.ProcessCombinedWithMultipleRounds(sortedDnsList, externalCoverageLines, filteredCache)
+            : ProcessingUtilities.ProcessCombined(sortedDnsList, externalCoverageLines, filteredCache);
+
+        if (settings.ExtraFiltering)
+        {
+            logger.LogInformation("Start extra filtering of duplicates");
+            sortedDnsList =
+                ProcessingUtilities.ProcessWithExtraFiltering(sortedDnsList, externalCoverageLines, filteredCache);
+            logger.LogInformation("Done extra filtering of duplicates");
+        }
+
+        await using StreamWriter streamWriter = new(settings.OutputFileName, false);
+        for (var i = 0; i < settings.HeaderLines.Length; i++)
+            await streamWriter.WriteLineAsync(settings.HeaderLines[i]);
+
+        await streamWriter.WriteLineAsync($"! Last Modified: {DateTime.UtcNow:u}");
+
+        foreach (var s in sortedDnsList)
+        {
+            await streamWriter.WriteLineAsync();
+            await streamWriter.WriteAsync((char) Constants.PipeSign);
+            await streamWriter.WriteAsync((char) Constants.PipeSign);
+            await streamWriter.WriteAsync(s);
+            await streamWriter.WriteAsync((char) Constants.HatSign);
+        }
+
+        stopWatch.Stop();
+        logger.LogInformation("Execution duration - {elapsed} | Produced {count} hosts", stopWatch.Elapsed,
+            sortedDnsList.Count);
+    }
+
+    private static async Task<(HashSet<string> CombinedLines, HashSet<string> ExternalCoverageLines)> ReadSources(Settings settings,
+        HttpClient httpClient)
+    {
+        var decoder = Encoding.UTF8.GetDecoder();
+        
+        // Assumed length to reduce allocations
+        var combineLines = new HashSet<string>(170_000);
+        var externalCoverageLines = new HashSet<string>(50_000);
+        for (var i = 0; i < settings.Filters.Sources.Length; i++)
+        {
+            var sourceItem = settings.Filters.Sources[i];
+            await using var stream = await httpClient.GetStreamAsync(sourceItem.Uri);
+            if (sourceItem.Format == SourceFormat.Hosts)
+            {
+                await HostUtilities.ProcessHostsBased(
+                    sourceItem.SourceAction == SourceAction.Combine ? combineLines : externalCoverageLines,
+                    stream,
+                    settings.Filters.SkipLinesBytes,
+                    sourceItem.SourcePrefixes,
+                    decoder);
+            }
+            else
+            {
+                await HostUtilities.ProcessAdBlockBased(
+                    sourceItem.SourceAction == SourceAction.Combine ? combineLines : externalCoverageLines,
+                    stream,
+                    decoder);
+            }
+        }
+
+        return (combineLines, externalCoverageLines);
+    }
+}
